# Projects/02-Backend Projects/003-Postaway Social Media API with MongoDB Mongoose/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"API for Social Media application",
        "title":"Social-Media API"
    },
    "servers":[
        {
            "url":"http://localhost:4200"
        }
     ],
     "components":{
        "securitySchemes":{
            "JWT":{
               "in":"header",
               "name":"Authorization",
               "type":"apiKey"
            }
     }
    },
     "paths":{
...
        "/api/posts":{
            "post":{
                "tags":["Posts"],
                "summary":"post details upload",
                "description":"upload post",
                "security":[{"JWT": {}}],
                "requestBody":{
                    "required":true,
                    "content":
                    {
                        "multipart/form-data": {
                            "schema": {
                              "type": "object",
                              "properties": {
                                "caption":{
                                    "type":"string"
                                },
                                "imageUrl":{
                                   
                                        "type": "string",
                                        "format": "binary"
                                     
                                }
                              }
                            }
                          }

                       
                       
                    }
                },
                "responses":{
                    "201":{
                        "description":"success"
                    }
                }
            },
            "get":{
                "tags":["Posts"],
                "summary":"get posts",
                "description":"get post details",
                "security":[{"JWT": {}}],
                "responses":{
                    "201":{
                        "description":"posts found"
                    },
                    "404":{
                        "description":"posts not found" 
                    }
                }
            }
        },
...
        "/api/posts/{postID}":{
            "get":{
                "tags":["Posts"],
                "summary":"get posts",
                "description":"get post details",
                "security":[{"JWT": {}}],
                "parameters":[{
                    "in":"path",
                    "name":"postID",
                    "required":true,
                    "schema":{
                        "type":"string"
                    }
                }],
                "responses":{
                    "201":{
                        "description":"posts found"
                    },
                    "404":{
                        "description":"posts not found" 
                    }
                }
            },
            "put":{
                "tags":["Posts"],
                "summary":"get posts",
                "description":"get post details",
                "security":[{"JWT": {}}],
                "parameters":[{
                    "in":"path",
                    "name":"postID",
                    "required":true,
                    "schema":{
                        "type":"string"
                    }
                }],
                "requestBody":{
                    "required":true,
                    "content":
                    {
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "properties":{
                                   
                                    "caption":{
                                        "type":"string"
                                    }
                                }
                            }
                        },
                        "image/*":{
                            "schema":{
                                "type":"string" 
                            }
                        }
                       
                    }
                },
                "responses":{
                    "201":{
                        "description":"posts found"
                    },
                    "404":{
                        "description":"post not found" 
                    }
                }
            },
            "delete":{
                "tags":["Posts"],
                "summary":"delete post by ID",
                "description":"delete post details",
                "security":[{"JWT": {}}],
                "parameters":[{
                    "in":"path",
                    "name":"postID",
                    "required":true,
                    "schema":{
                        "type":"string"
                    }
                }],
                "responses":{
                    "201":{
                        "description":"post deleted"
                    }
                }
            }
        },
...
    }
    }